Remove package checking functionality
Now that most of LXC/LXD is installed through Snap, checking the packages with the apt cache doesn't make much sense anymore.
Also, the README clearly states which packages are required, so checking on every command is a bit overkill.

Also, update Black to the latest version to support py3.10